Split package into basis package and ecpackage
